requirements.txt
llama-index
llama-index-multi-modal-llms-gemini
llama-index-vector-stores-chroma
llama-index-embeddings-gemini
llama-index-embeddings-huggingface
python-dotenv
streamlit
pandas
pillow
llama-parse
